% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{bansal_2021_metadata}
\alias{bansal_2021_metadata}
\title{Isolate metadata from Bansal et al. 2021}
\format{
\code{data.frame}
}
\source{
Bansal, K., Kumar, S., Kaur, A., Singh, A. and Patil, P.B., 2021.
Deep phylo-taxono genomics reveals Xylella as a variant lineage of plant
associated Xanthomonas and supports their taxonomic reunification along
with Stenotrophomonas and Pseudoxanthomonas. Genomics, 113(6),
pp.3989-4003.
}
\description{
Isolate metadata from Bansal et al. 2021
(\doi{10.1016/j.ygeno.2021.09.021}) included for use as an example
data set. This is the parsed version of the included "bansal_2021_metadata.tsv"
file.
}
\keyword{data}
