% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{example_metadata_1}
\alias{example_metadata_1}
\title{Example metadata for the tips of \code{example_tree_1}}
\format{
\code{data.frame}
}
\source{
From a core gene phylogeny of publicly available sequences
}
\description{
The parsed version of the included "example_metadata_1".
}
\keyword{data}
