% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{example_tree_1}
\alias{example_tree_1}
\title{An example tree in "phylo" format}
\format{
\code{ape::phylo}
}
\source{
From a core gene phylogeny of publicly available sequences
}
\description{
The parsed version of the included "example_tree_1.treefile".
}
\keyword{data}
