% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_tree.R
\name{heat_tree}
\alias{heat_tree}
\title{Create an interactive phylogenetic tree}
\usage{
heat_tree(
  tree = NULL,
  metadata = NULL,
  aesthetics = NULL,
  width = NULL,
  height = NULL,
  elementId = NULL,
  ...
)
}
\arguments{
\item{tree}{One or more trees to plot. Can be a raw newick-formatted string,
a \code{phylo} object, or a list of such inputs. If a list is provided, its
names will be used to name trees.}

\item{metadata}{Metadata associated with \code{tree}. Can be a \code{data.frame} or
\code{tibble} with a "node_id" column. If there are multiple trees (a list),
then a list of table of equal length is required. If a list is provided,
its names will be used to name metadata tables.}

\item{aesthetics}{A named character vector defining which metadata columns
are initially used to color/size tree parts. If there are multiple trees (a
list), then a list of equal length is required.}

\item{width}{Width of the widget (CSS units or number).}

\item{height}{Height of the widget (CSS units or number).}

\item{elementId}{Optional element ID for the widget.}

\item{...}{Options passed to options parameter of the underlying \code{HeatTree.heatTree()} javascript
function to modify the initial state of the widget.}
}
\description{
Create an interactive phylogenetic tree using the javascript \href{https://github.com/grunwaldlab/heat-tree}{heat-tree} package
}
