% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{weisberg_2020_beast}
\alias{weisberg_2020_beast}
\title{A BEAST phylogeny from Weisberg et al. 2020}
\format{
\code{ape::phylo}
}
\source{
Alexandra J. Weisberg et al., Unexpected conservation and global
transmission of agrobacterial virulence plasmids.Science368,eaba5256(2020)
}
\description{
A BEAST phylogeny from Weisberg et al. 2020
(\doi{10.1126/science.aba5256}) included for use as an example
data set. This is the parsed version of the included "weisberg_2020_beast.tre"
file.
}
\keyword{data}
